==> notebooks/01_load_and_validate_raw_data.ipynb <==
# 📕 01 - Load and Validate Raw Data

## Introduction 

This notebook focuses on the data ingestion and pre-processing of the raw rides dataset. The data is sourced from TLC and contain detailed trip data from taxi rides for a specific month. Our goal is to download, explore, validate, and clean this data for further analysis and modeling.

## TLC Trip Record Data

[The New York City Taxi and Limousine Commission (TLC)](https://www.nyc.gov/site/tlc/about/tlc-trip-record-data.page) provides access to extensive trip records from both yellow and green taxis as well as For-Hire Vehicles (FHV). This data is invaluable for a variety of applications, including urban planning, traffic modeling, and socio-economic studies.

### Yellow and Green Taxi Trip Records

These records offer a detailed look into the operations of traditional taxis in the city:

- **Date/Time**: Each record contains timestamps for both the pick-up and drop-off events.
- **Locations**: Precise locations, both for pick-up and drop-off, are included. 
- **Trip Details**: Data encompasses the trip distance, a breakdown of the fare, the rate type, payment methods, and the number of passengers as reported by the driver.
- **Source**: The data stems from technology providers sanctioned under the Taxicab & Livery Passenger Enhancement Programs (TPEP/LPEP). It's worth noting that while the 
TLC has provided access to this data, they did not generate it and thus do not guarantee its accuracy.

## Notebook Overview:

1. **Data Download**:  I've crafted a function that streamlines the download process based on the specified month and year. For the initial phase of this project, I've chosen an arbitrary date and will be working with data from a single month.
2. **Data Exploration**: A swift perusal of the basic data structure and principal columns we're interested in.
3. **Data Cleaning**: 
    - Filtering relevant columns.
    - Renaming columns for clarity.
    - Ensuring the data is within the desired date range.
4. **Data Saving**: The validated and cleaned data is then saved into a new Parquet file for further usage.

# import necessary libraries
from pathlib import Path
import requests
import pandas as pd

def download_one_file(year: int, month: int) -> Path:
    """
    Downloads the yellow taxi trip data for a given year and month from the TLC website and saves it as a parquet file.

    Args:
    - year (int): The year of the data to download.
    - month (int): The month of the data to download.

    Returns:
    - path (Path): The path to the downloaded parquet file.
    """

    URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'
    response = requests.get(URL)

    if response.status_code == 200:
        path = f'../data/raw/rides_{year}-{month:02d}.parquet'
        open(path, 'wb').write(response.content)
        return path
    else:
        raise Exception(f'{URL} is not available.')

# explore and validate one single file
download_one_file(year=2022, month=1)

rides = pd.read_parquet('../data/raw/rides_2022-01.parquet')
rides.head(20)

rides = rides[['tpep_pickup_datetime', 'PULocationID']]
rides

# rename columns to a more convenient format
rides.rename(columns={
    'tpep_pickup_datetime': 'pickup_datetime',
    'PULocationID': 'pickup_location_id'
}, inplace=True)

# check data description
rides.describe()

# remove rides out of the time range
rides = rides[rides['pickup_datetime'] < '2022-02-01']
rides = rides[rides['pickup_datetime'] > '2022-01-01']

# check again for data description
rides.pickup_datetime.describe()

# export validated data to parquet file
rides.to_parquet('../data/transformed/validated_rides_2022_01.parquet')
