==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
# 📕 02 - Transform Raw Data into Time Series Data

## Introduction

In the preceding notebook, `01_load_and_validate_raw_data.ipynb`, we embarked on the initial journey of ingesting and pre-processing the TLC's raw taxi ride data. With the refined dataset in hand, our next mission is to shape this information into a structured time-series format.

The prime objective? To aggregate the number of rides according to two pivotal axes: `pickup_hour` and `pickup_location_id`, ensure continuity by addressing potential gaps in houw dataset.

## Notebook Overview:

1. **Validated Dataset Retrieval**: Begin by importing the validated dataset.
2. **Hourly Binning**: Isolate the exact timestamps of each ride and group them into hourly segments. This approach simplifies our dataset by representing each ride by the hour it commenced, rather than its precise starting moment. 
3. **Aggregation**: Compile the data by both hour and location, yielding a concise count of rides for every unique `pickup_hour`-`pickup_location_id` combination.
4. **Ensuring Continuity**: Time-series data thrives on regularity. Address potential gaps in our dataset by ensuring every location possesses entries for the entire temporal span.
5. **Visualization**: To validate and explore our time-series data further, draw a dynamic line chart depicting the frequency of rides over time for select pickup locations.

# import libraries
import pandas as pd
from tqdm import tqdm
from typing import Optional, List
import plotly.express as px

# Load the validated raw data and check the first 20 entries
rides = pd.read_parquet('../data/transformed/validated_rides_2022_01.parquet')
rides.head(20)

# Here, we transform the precise `pickup_datetime` into hourly intervals
rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('H')
rides

# aggregate the data
agg_rides = rides.groupby(['pickup_hour', 'pickup_location_id']).size().reset_index()
agg_rides.rename(columns={0: 'rides'}, inplace=True)

agg_rides


def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    """
    Add missing hourly slots to a DataFrame of ride counts, filling them with zeros.

    Parameters:
    - agg_rides (pd.DataFrame): A DataFrame containing ride counts, with columns 'pickup_hour', 'pickup_location_id', and 'rides'.

    Returns:
    - pd.DataFrame: A new DataFrame with the same columns as `agg_rides`, but with additional rows for any missing hourly slots, filled with zeros.

    Example:
    >>> rides = pd.DataFrame({'pickup_hour': ['2022-01-01 00:00:00', '2022-01-01 02:00:00'], 'pickup_location_id': [1, 1], 'rides': [10, 20]})
    >>> add_missing_slots(rides)
    pickup_hour         pickup_location_id   rides
    2022-01-01 00:00:00 1                    10
    2022-01-01 01:00:00 1                    0
    2022-01-01 02:00:00 1                    20
    """
    
    location_ids = agg_rides['pickup_location_id'].unique()
    full_range = pd.date_range(agg_rides['pickup_hour'].min(), agg_rides['pickup_hour'].max(), freq='H')
    output = pd.DataFrame()

    for location_id in tqdm(location_ids):
        agg_rides_i = agg_rides.loc[agg_rides['pickup_location_id'] == location_id, ['pickup_hour', 'rides']]
        agg_rides_i.set_index('pickup_hour', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)
        agg_rides_i['pickup_location_id'] = location_id

        output = pd.concat([output, agg_rides_i])

    output = output.reset_index().rename(columns={'index': 'pickup_hour'})

    return output


agg_rides_all_slots = add_missing_slots(agg_rides)

def plot_rides(
        rides: pd.DataFrame,
        locations: Optional[List[int]] = None
) -> None:
    """
    Plot a line chart of the number of rides over time for specified pickup locations using Plotly.

    Parameters:
    - rides (pd.DataFrame): DataFrame containing ride information. It must have columns 'pickup_hour', 'rides', and 'pickup_location_id'.
    - locations (Optional[List[int]], default=None): A list of pickup_location_id values to be plotted. 
        If None, all locations in the 'rides' DataFrame will be plotted.

    Returns:
    None: Shows the plot using Plotly.
    
    Example:
    >>> plot_rides(df_rides, locations=[123, 456])
    """
    
    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    fig = px.line(
        rides_to_plot,
        x='pickup_hour',
        y='rides',
        color='pickup_location_id',
        template='none'
    )

    fig.show()

# plot the rides for locations 4 and 43
plot_rides(agg_rides_all_slots, locations=[4, 43])


# save aggregated data to parquet
agg_rides_all_slots.to_parquet("../data/transformed/ts_data_rides_2022_01.parquet")
