==> notebooks/03_transform_ts_data_into_features_and_targets.ipynb <==
# 📕 03 - Transform Time Series Data into Features and Targets

## Introduction

Navigating from the previous steps where we structured the raw data into a time-series format, our next leap in the data processing journey is to carve out discernible features and targets from our aggregated dataset. This transformation paves the way for our upcoming machine learning endeavors, ensuring we have a well-defined problem space and solution mechanism. 

## Notebook Overview:

1. **Data Subset**: Extract data for a specific location.
2. **Utility Development**: Create a function to identify cutoff indices.
3. **Indexing Strategy**: Define the start, middle, and end indices for each sub-sequence.
4. **Data Transformation**: Convert time series data into features and targets.



# import libraries
import pandas as pd
import numpy as np
from tqdm import tqdm
from typing import Tuple

# load ts data nd check first rows
ts_data = pd.read_parquet('../data/transformed/ts_data_rides_2022_01.parquet')
ts_data

# Subset the data to select entries for a specific pickup location
ts_data_one_location = ts_data.loc[ts_data['pickup_location_id'] == 43, :].reset_index(drop=True)
ts_data_one_location.head(15)

def get_cutoff_indices(
        data: pd.DataFrame,
        n_features: int,
        step_size: int
) -> list:
    """
    Given a time series data and the number of features and step size, returns a list of tuples containing the start,
    middle and end indices of each sub-sequence.

    :param data: A pandas DataFrame containing the time series data.
    :param n_features: An integer representing the number of features to use in each sub-sequence.
    :param step_size: An integer representing the step size between each sub-sequence.
    :return: A list of tuples containing the start, middle and end indices of each sub-sequence.
    """
    stop_position = len(data) - 1

    # start the first sub-sequence at index position 0
    subseq_first_idx = 0
    subseq_mid_idx = n_features
    subseq_last_idx = n_features + 1
    indices = []

    while subseq_last_idx <= stop_position:
        indices.append((subseq_first_idx, subseq_mid_idx, subseq_last_idx))

        subseq_first_idx += step_size
        subseq_mid_idx += step_size
        subseq_last_idx += step_size

    return indices


# Setting parameters for feature extraction
n_features = 24
step_size = 1

# Compute indices for data split
indices = get_cutoff_indices(ts_data_one_location, n_features, step_size)
n_examples = len(indices)

indices[:5]

# Initialize arrays to store the features and targets
x = np.ndarray(shape=(n_examples, n_features), dtype=np.float32)
y = np.ndarray(shape=(n_examples), dtype=np.float32)
pickup_hours = []

# Convert time series data to features and targets for the selected location
for i, idx in enumerate(indices):
    x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
    y[i] = ts_data_one_location.iloc[idx[1]]['rides']
    pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

# Display shape and values of the features and targets for validation
print(f"{x.shape=}")
print(f"{x=}")
print(f"{y=}")
print(f"{pickup_hours[:5]=}")

# Create a dataframe for the features
features_one_location = pd.DataFrame(x,
    columns=[f'rides_previous_{i+1}_hour' for i in reversed(range(n_features))])

features_one_location

# Create a dataframe for the targets
targets_one_location = pd.DataFrame(y,
    columns=['target_rides_next_hour'])

targets_one_location

For our taxi demand prediction task, we transform our time series dataset into features and targets suitable for training machine learning models. This involves creating sequences or "windows" of historical data as features and the subsequent hour's data as the target.

Parameters used:
- **`ts_data`**: The entire time series dataset.
- **`n_features`**: This specifies the length of the historical window we are considering. A value of `24*7` means we are using the past week (7 days, with 24 hours each) of data to predict the taxi demand of the subsequent hour.
- **`step_size`**: The number of hours the window moves forward for each subsequent data point. A value of `24` means we move forward by a day for each step, creating a new feature set for every day in our dataset.

The output consists of two parts:
- **`features`**: This contains the sequences of historical data. Each row represents a sequence of the past week's taxi demand data, and each column in that row corresponds to an hour's data.
- **`targets`**: This contains the actual taxi demand values for the hour immediately following each historical sequence in `features`.


def transform_ts_data_into_features_and_target(
    ts_data: pd.DataFrame,
    n_features: int,
    step_size: int
) -> Tuple[pd.DataFrame, pd.Series]:
    """
    Transforms a time series dataset into features and targets suitable for machine learning.

    Given a time series dataset with columns 'pickup_hour', 'rides' and 'pickup_location_id', this function
    extracts the data for each location, and for each location it creates a set of features and targets. The
    features are created by selecting a window of n_features hours of rides data, and the target is the number
    of rides in the next hour. The window is moved step_size hours at a time, creating multiple examples. The
    function returns two dataframes: one with the features, and one with the targets.

    :param ts_data: A pandas DataFrame containing the time series data.
    :param n_features: An integer representing the number of features to use in each sub-sequence.
    :param step_size: An integer representing the step size between each sub-sequence.
    :return: A tuple containing two pandas DataFrames: the first one with the features, and the second one with
            the targets.
    """
    # check that the ts_data DataFrame has the expected columns
    assert set(ts_data.columns) == {'pickup_hour', 'rides', 'pickup_location_id'}

    # get the unique location ids
    location_ids = ts_data["pickup_location_id"].unique()
    features = pd.DataFrame()
    targets = pd.DataFrame()

    for location_id in tqdm(location_ids):
        ts_data_one_location = ts_data.loc[ts_data['pickup_location_id'] == location_id, ['pickup_hour', 'rides']]

        # pre-compute indices for one location
        indices = get_cutoff_indices(ts_data_one_location, n_features, step_size)
        n_examples = len(indices)
        
        # transform time series data into features and targets
        x = np.ndarray(shape=(n_examples, n_features), dtype=np.float32)
        y = np.ndarray(shape=(n_examples))
        pickup_hours = []

        for i, idx in enumerate(indices):
            x[i] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
            y[i] = ts_data_one_location.iloc[idx[2]]['rides']
            pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

        features_one_location = pd.DataFrame(x, columns=[f'rides_previous_{i+1}_hour' for i in reversed(range(n_features))])
        
        features_one_location['pickup_hour'] = pickup_hours
        features_one_location['pickup_location_id'] = location_id

        targets_one_location = pd.DataFrame(y, columns=["target_rides_next_hour"])

        # concatenate features and targets for one location
        features = pd.concat([features, features_one_location])
        targets = pd.concat([targets, targets_one_location])

    features.reset_index(inplace=True, drop=True)
    targets.reset_index(inplace=True, drop=True)

    return features, targets['target_rides_next_hour']




# Apply the transformation to the entire dataset
features, targets = transform_ts_data_into_features_and_target(
    ts_data=ts_data,
    n_features=24*7*1,
    step_size=1
)

# Display shapes for validation
print(f'{features.shape=}')
print(f'{targets.shape=}')

# Display the first few rows of the transformed features
features.head(24)
