==> notebooks/05_plot_training_data.ipynb <==
# 📕 05 - Plot Training Data

## Introduction

After going through rigorous processes of data collection, validation, and transformation in our previous notebooks, we now take a last visual approach. The primary purposes of this exploration are:

1. **Insight Discovery:** Charts and graphs can reveal patterns, outliers, or anomalies in the data that might not be immediately evident from raw tables or basic statistics. They can guide us in understanding the nuances of our dataset, which can be immensely valuable when building and refining our models later.

2. **Validation & Quality Assurance:** This is our last step before we start model building. It's crucial to ensure that our transformations and processing steps have not introduced errors. Mistakes can happen, and it's much easier to catch these visually than by looking at rows of numbers. This step is a "last line of defense" against potential issues in our dataset.

# import libraries
import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR
from src.plot import plot_one_sample

# load features and target data
features_and_target = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
features_and_target.head(20)

# split into features and target
features = features_and_target.drop(columns=['target_rides_next_hour'])
targets = features_and_target['target_rides_next_hour']

# plot one sample
plot_one_sample(
    example_id=24261,
    features=features,
    targets=targets,
)
