==> notebooks/10_lightgbm_hyperparameter_tunning.ipynb <==
import warnings
warnings.filterwarnings("ignore")

# import libraries
from datetime import datetime
import pandas as pd
import optuna
import numpy as np
from sklearn.model_selection import TimeSeriesSplit
from sklearn.metrics import mean_absolute_error
from src.paths import TRANSFORMED_DATA_DIR
from src.data_split import train_test_split
from src.model import get_pipeline

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
df.head()

df.pickup_hour.max()

X_train, y_train, X_test, y_test = train_test_split(
    df=df,
    cutoff_date=datetime(2022, 8, 1, 0, 0),
    target_col_name="target_rides_next_hour",
)

# print shapes of train and test data
print("X_train shape:", X_train.shape)
print("y_train shape:", y_train.shape)
print("X_test shape:", X_test.shape)
print("y_test shape:", y_test.shape)

def create_objective(X_train, y_train):
    def objective(trial: optuna.trial.Trial) -> float:
        """
        Train a model with a set of hyperparameters and compute and validation error based on a TimeSeriesSplit.
        """
        # define hyperparameters to tune
        hyperparams = {
            "metric": "mae",
            "verbose": -1,
            "num_leaves": trial.suggest_int("num_leaves", 2, 256),
            "feature_fraction": trial.suggest_float("feature_fraction", 0.2, 1.0),
            "bagging_fraction": trial.suggest_float("bagging_fraction", 0.2, 1.0),
            "min_child_samples": trial.suggest_int("min_child_samples", 3, 100),    
        }

        tss = TimeSeriesSplit(n_splits=2)
        scores = []
        for train_index, val_index in tss.split(X_train):
            X_train_, X_val_ = X_train.iloc[train_index], X_train.iloc[val_index]
            y_train_, y_val_ = y_train.iloc[train_index], y_train.iloc[val_index]

            # create pipeline
            pipeline = get_pipeline(**hyperparams)
            pipeline.fit(X_train_, y_train_)

            # evaluate the model
            y_pred = pipeline.predict(X_val_)
            mae = mean_absolute_error(y_val_, y_pred)
            scores.append(mae)

        # return the mean scores
        return np.array(scores).mean()

    return objective

study = optuna.create_study(direction="minimize")
study.optimize(create_objective(X_train, y_train), n_trials=5)

best_params = study.best_trial.params
print(f'{best_params=}')

pipeline = get_pipeline(**best_params)
pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
