==> notebooks/11_backfill_feature_store.ipynb <==
# reload kernel after running this script
%load_ext autoreload
%autoreload 2

# import libraries
import os
from datetime import datetime

import pandas as pd
import hopsworks
from dotenv import load_dotenv

from src.paths import PARENT_DIR
from src.data import load_raw_data
from src.data import transform_raw_data_into_ts_data

load_dotenv(PARENT_DIR / ".env")

HOPSWORKS_API_KEY = os.environ["HOPSWORKS_API_KEY"]
HOPSWORKS_PROJECT_NAME = 'taxi_demand_api'

start_year = 2022
end_year = datetime.now().year

print(f'Loading data from {start_year} to {end_year}...')

rides = pd.DataFrame()
for year in range(start_year, end_year + 1):
    one_year_rides = load_raw_data(year)
    rides = pd.concat([rides, one_year_rides])

print(f'{len(rides)=}')

ts_data = transform_raw_data_into_ts_data(rides)

# establish connection to Hopsworks
project = hopsworks.login(
    project=HOPSWORKS_PROJECT_NAME,
    api_key_value=HOPSWORKS_API_KEY,
)

# get feature store handle
feature_store = project.get_feature_store()

FEATURE_GROUP_NAME = 'time_series_hourly_feature_group'
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description='Time series data aggregated by hour',
    primary_key=['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour',
)

feature_group.insert(ts_data, write_options={"wait_for_job": False})

ts_data.shape

rides.shape
