==> notebooks/14_inference_pipeline.ipynb <==
%load_ext autoreload
%autoreload 2

# Standard Library Imports
from datetime import datetime, timedelta

# Third-party imports
import pandas as pd

# Local module imports
from src.inference import load_batch_of_features_from_store, load_model_from_registry, get_model_predictions
from src.feature_store_api import get_feature_store
import src.config as config


current_date = pd.to_datetime(datetime.utcnow()).floor('H')

features = load_batch_of_features_from_store(current_date)

features.head()

model = load_model_from_registry()
predictions = get_model_predictions(model, features)

predictions['pickup_hour'] = current_date
predictions

feature_group = get_feature_store().get_or_create_feature_group(
    name=config.FEATURE_GROUP_MODEL_PREDICTIONS,
    version=1,
    description="Predictions generated by the production model",
    primary_key=["pickup_location_id", "pickup_hour"],
    event_time="pickup_hour"
    )

feature_group.insert(predictions, write_options={"wait_for_job": False})


predictions
